feat: Hide experiments from archived projects in `GetExperiments` Added a condition that filters sql results when parent workspace or project is archived and wrote a test to verify Adding more tests and resolving comments
